+++ sapp/forms.py
@@ -7,10 +7,10 @@ from dataclasses import replace
 import npyscreen
 from slash import Slash
 
-from . import slurm_utils as utils
+from . import utils as utils
 from .core import Database
 from .gpustat import get_card_list
-from .slurm_config import SlurmConfig, SubmitConfig
+from .config import SlurmConfig, SubmitConfig
 
 
 def satisfy(req: dict, avail: dict) -> int:
